refactor: broke entities into separate files
