+++ packages/deployment-service/src/entities/abstract-entity.ts
@@ -1,0 +1,12 @@
+import { CreateDateColumn, PrimaryGeneratedColumn, UpdateDateColumn } from 'typeorm'
+
+export abstract class AbstractEntity {
+  @PrimaryGeneratedColumn('uuid')
+  id?: string
+
+  @CreateDateColumn()
+  created?: string
+
+  @UpdateDateColumn()
+  modified?: string
+}
